# packages/cactus-plugin-ledger-connector-ethereum/src/main/json/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hyperledger Cacti Plugin - Connector Ethereum",
    "description": "Can perform basic tasks on a Ethereum ledger",
    "version": "0.0.1",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "components": {
    "schemas": {
      "Web3SigningCredential": {
        "type": "object",
        "required": ["type"],
        "discriminator": {
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/Web3SigningCredentialGethKeychainPassword"
          },
          {
            "$ref": "#/components/schemas/Web3SigningCredentialCactiKeychainRef"
          },
          {
            "$ref": "#/components/schemas/Web3SigningCredentialPrivateKeyHex"
          },
          {
            "$ref": "#/components/schemas/Web3SigningCredentialNone"
          }
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/Web3SigningCredentialType"
          }
        }
      },
      "Web3SigningCredentialGethKeychainPassword": {
        "type": "object",
        "required": ["type", "ethAccount", "secret"],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/Web3SigningCredentialType"
          },
          "ethAccount": {
            "type": "string",
            "description": "The ethereum account (public key) that the credential  belongs to. Basically the username in the traditional terminology of authentication.",
            "minLength": 66,
            "maxLength": 66,
            "nullable": false
          },
          "secret": {
            "type": "string",
            "description": "A geth keychain unlock password.",
            "minLength": 0,
            "maxLength": 65535
          }
        }
      },
      "Web3SigningCredentialCactiKeychainRef": {
        "type": "object",
        "required": ["type", "ethAccount", "keychainEntryKey"],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/Web3SigningCredentialType"
          },
          "ethAccount": {
            "type": "string",
            "description": "The ethereum account (public key) that the credential  belongs to. Basically the username in the traditional  terminology of authentication.",
            "minLength": 66,
            "maxLength": 66,
            "nullable": false
          },
          "keychainEntryKey": {
            "type": "string",
            "description": "The key to use when looking up the the keychain entry holding the secret pointed to by the  keychainEntryKey parameter.",
            "minLength": 0,
            "maxLength": 1024
          },
          "keychainId": {
            "type": "string",
            "description": "The keychain ID to use when looking up the the keychain plugin instance that will be used to retrieve the secret pointed to by the keychainEntryKey parameter.",
            "minLength": 0,
            "maxLength": 1024
          }
        }
      },
      "Web3SigningCredentialPrivateKeyHex": {
        "type": "object",
        "required": ["type", "ethAccount", "secret"],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/Web3SigningCredentialType"
          },
          "ethAccount": {
            "type": "string",
            "description": "The ethereum account (public key) that the credential belongs to. Basically the username in the traditional terminology of authentication.",
            "minLength": 66,
            "maxLength": 66,
            "nullable": false
          },
          "secret": {
            "type": "string",
            "description": "The HEX encoded private key of an eth account.",
            "minLength": 0,
            "maxLength": 65535
          }
        }
      },
      "Web3SigningCredentialNone": {
        "type": "object",
        "required": ["type"],
        "description": "Using this denotes that there is no signing required because the transaction is pre-signed.",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/Web3SigningCredentialType"
          }
        }
      },
      "Web3SigningCredentialType": {
        "type": "string",
        "enum": [
          "CACTI_KEYCHAIN_REF",
          "GETH_KEYCHAIN_PASSWORD",
          "PRIVATE_KEY_HEX",
          "NONE"
        ]
      },
...
      "GasTransactionConfigLegacy": {
        "type": "object",
        "description": "Transaction gas settings in networks before EIP-1559 (London fork).",
        "required": [],
        "properties": {
          "gas": {
            "type": "string",
            "description": "A maximum amount of gas a user is willing to provide for the execution of the transaction. (gasLimit)"
          },
          "gasPrice": {
            "type": "string",
            "description": "A price (in Wei) a user is willing to pay for each unit of gas used during the execution of the transaction. In EIP-1559 (London fork) networks, it will be set as both maxFeePerGas and maxPriorityFeePerGas."
          }
        }
      },
      "GasTransactionConfigEIP1559": {
        "type": "object",
        "description": "Transaction gas settings in networks after EIP-1559 (London fork).",
        "required": [],
        "properties": {
          "gasLimit": {
            "type": "string",
            "description": "A maximum amount of gas a user is willing to provide for the execution of the transaction.",
            "nullable": false
          },
          "maxFeePerGas": {
            "type": "string",
            "description": "A maximum fee (including the base fee and the tip) a user is willing to pay per unit of gas.",
            "nullable": false
          },
          "maxPriorityFeePerGas": {
            "type": "string",
            "description": "A maximum tip amount a user is willing to pay per unit of gas.",
            "nullable": false
          }
        }
      },
      "GasTransactionConfig": {
        "type": "object",
        "description": "Transaction gas settings.",
        "required": [],
        "oneOf": [
          {
            "$ref": "#/components/schemas/GasTransactionConfigLegacy"
          },
          {
            "$ref": "#/components/schemas/GasTransactionConfigEIP1559"
          }
        ]
      },
...
      "Web3TransactionReceipt": {
        "type": "object",
        "required": [
          "blockHash",
          "blockNumber",
          "transactionHash",
          "transactionIndex",
          "status",
          "from",
          "to",
          "gasUsed"
        ],
        "additionalProperties": true,
        "properties": {
          "status": {
            "type": "boolean",
            "nullable": false
          },
          "transactionHash": {
            "type": "string",
            "minLength": 66,
            "maxLength": 66,
            "pattern": "^0x([A-Fa-f0-9]{64})$"
          },
          "transactionIndex": {
            "type": "string",
            "nullable": false
          },
          "blockHash": {
            "type": "string",
            "minLength": 66,
            "maxLength": 66,
            "pattern": "^0x([A-Fa-f0-9]{64})$"
          },
          "blockNumber": {
            "type": "string",
            "nullable": false
          },
          "gasUsed": {
            "type": "string",
            "nullable": false
          },
          "effectiveGasPrice": {
            "type": "string",
            "nullable": false
          },
          "contractAddress": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": false
          },
          "to": {
            "type": "string",
            "nullable": false
          },
          "logs": {
            "type": "array",
            "default": [],
            "items": {},
            "nullable": false
          },
          "logsBloom": {
            "type": "string",
            "nullable": false
          },
          "revertReason": {
            "type": "string",
            "nullable": false
          },
          "output": {
            "type": "string",
            "nullable": false
          },
          "commitmentHash": {
            "type": "string",
            "nullable": false
          },
          "cumulativeGasUsed": {
            "type": "string",
            "nullable": false
          },
          "type": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "ContractJSON": {
        "type": "object",
        "required": ["contractName", "bytecode", "abi"],
        "additionalProperties": true,
        "properties": {
          "contractName": {
            "type": "string",
            "nullable": false
          },
          "bytecode": {
            "type": "string",
            "nullable": false,
            "minLength": 1,
            "maxLength": 24576,
            "description": "See https://ethereum.stackexchange.com/a/47556 regarding the maximum length of the bytecode"
          },
          "abi": {
            "description": "The application binary interface of the solidity contract, optional parameter",
            "type": "array",
            "items": {},
            "nullable": false
          },
          "metadata": {
            "type": "string"
          },
          "deployedBytecode": {
            "type": "string"
          },
          "sourceMap": {
            "type": "string"
          },
          "deployedSourceMap": {
            "type": "string"
          },
          "sourcePath": {
            "type": "string"
          },
          "compiler": {
            "type": "object"
          },
          "networks": {
            "type": "object"
          },
          "ast": {
            "type": "object"
          },
          "functionHashes": {
            "type": "object"
          },
          "gasEstimates": {
            "type": "object"
          }
        }
      },
      "ContractJsonDefinition": {
        "type": "object",
        "required": ["contractJSON"],
        "additionalProperties": false,
        "properties": {
          "contractJSON": {
            "$ref": "#/components/schemas/ContractJSON",
            "nullable": false
          }
        }
      },
...
      "ContractKeychainDefinition": {
        "type": "object",
        "required": ["contractName", "keychainId"],
        "additionalProperties": false,
        "properties": {
          "contractName": {
            "type": "string",
            "description": "The contract name for retrieve the contracts json on the keychain.",
            "minLength": 1,
            "maxLength": 100,
            "nullable": false
          },
          "keychainId": {
            "type": "string",
            "description": "The keychainId for retrieve the contracts json.",
            "minLength": 1,
            "maxLength": 100,
            "nullable": false
          }
        }
      },
...
      "RunTransactionResponse": {
        "type": "object",
        "required": ["transactionReceipt"],
        "properties": {
          "transactionReceipt": {
            "$ref": "#/components/schemas/Web3TransactionReceipt"
          }
        }
      },
      "DeployContractV1Request": {
        "type": "object",
        "required": ["web3SigningCredential", "contract"],
        "additionalProperties": false,
        "properties": {
          "web3SigningCredential": {
            "$ref": "#/components/schemas/Web3SigningCredential",
            "nullable": false
          },
          "contract": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ContractJsonDefinition",
                "description": "Send contract ABI directly in the request."
              },
              {
                "$ref": "#/components/schemas/ContractKeychainDefinition",
                "description": "Read contract definition from the keychain plugin."
              }
            ],
            "nullable": false
          },
          "constructorArgs": {
            "description": "The list of arguments to pass in to the constructor of the contract being deployed.",
            "type": "array",
            "default": [],
            "items": {}
          },
          "gasConfig": {
            "$ref": "#/components/schemas/GasTransactionConfig",
            "nullable": false
          },
          "value": {
            "type": "string",
            "description": "Ether balance to send on deployment.",
            "nullable": false
          }
        }
      },
...
      "ErrorExceptionResponseV1": {
        "type": "object",
        "description": "Error response from the connector.",
        "required": ["message", "error"],
        "properties": {
          "message": {
            "type": "string",
            "description": "Short error description message.",
            "nullable": false
          },
          "error": {
            "type": "string",
            "description": "Detailed error information.",
            "nullable": false
          }
        }
      }
    }
  },
  "paths": {
    "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-ethereum/deploy-contract": {
      "post": {
        "x-hyperledger-cacti": {
          "http": {
            "verbLowerCase": "post",
            "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-ethereum/deploy-contract"
          }
        },
        "operationId": "deployContract",
        "summary": "Deploys the contract to ethereum ledger.",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployContractV1Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunTransactionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorExceptionResponseV1"
                }
              }
            }
          }
        }
      }
    },
...
  }
}